Split `integration.rs`
