+++ tests/utils/command.rs
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: 2024 Shun Sakai
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+pub fn command() -> assert_cmd::Command {
+    let mut command = assert_cmd::Command::cargo_bin(env!("CARGO_PKG_NAME")).unwrap();
+    command.current_dir("tests");
+    command
+}
